Added Cucumber tests for the Currency enum Updated the Currency enum to handle BigDecimal instead of double

+++ src/main/java/pro/tremblay/roi/service/ExchangeRateService.java
@@ -38,6 +38,6 @@ public class ExchangeRateService extends DependencyService {
      * @return change rate
      */
     public BigDecimal getExchangeRate(Currency origin, Currency destination, LocalDate date) {
-        return BigDecimal.valueOf(origin.convertTo(destination));
+        return origin.convertTo(destination);
     }
 }
